+++ pkg/extension/sumologicextension/factory_test.go
@@ -77,19 +77,3 @@ func TestFactory_CreateExtension(t *testing.T) {
 	require.NoError(t, err)
 	require.NotNil(t, ext)
 }
-
-func TestFactory_CreateExtensionWithInstallToken(t *testing.T) {
-	cfg := createDefaultConfig().(*Config)
-	cfg.CollectorName = "test_collector"
-	cfg.Credentials.InstallToken = "dummy_install_token"
-
-	ext, err := createExtension(context.Background(),
-		extension.CreateSettings{
-			TelemetrySettings: componenttest.NewNopTelemetrySettings(),
-		},
-		cfg,
-	)
-	require.NoError(t, err)
-	require.NotNil(t, ext)
-	require.Equal(t, configopaque.String("dummy_install_token"), cfg.Credentials.InstallationToken)
-}