+++ CHANGELOG.md
@@ -9,6 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Released TBD
 
+### Changed
+
+- feat(sumologicextension)!: remove support for `install_token` [#1225]
+
+[#1225]: https://github.com/SumoLogic/sumologic-otel-collector/pull/1225
 [Unreleased]: https://github.com/SumoLogic/sumologic-otel-collector/compare/v0.82.0-sumo-0...main
 
 ## [v0.82.0-sumo-0]